requirements.txt
matplotlib==3.0.2
scipy==1.2.1
tqdm==4.31.1
opencv_python==4.1.1.26
tensorflow==1.15.0rc2
networkx==2.2
numpy==1.18.4
moviepy==1.0.1
trimesh==2.37.35
tensorflow_plot==0.3.0
scikit_image==0.15.0
colormap==1.0.2
maya==0.6.1
PyOpenGL==3.1.5
PySide2==5.15.0
PyYAML==5.3.1
scikit_learn==0.23.1
skimage==0.0
